✨ feat: Implement accountant and tracking modes in group detail balances
