+++ src/features/group-detail/balances/TrackingMode.vue
@@ -1,0 +1,5 @@
+<script setup lang="ts"></script>
+
+<template>
+	<div>Hello</div>
+</template>
