+++ src/features/group-detail/hooks/useGroupContext.ts
@@ -1,9 +1,12 @@
 import { CONTEXT_KEY } from '@/constants/key';
-import type { Group } from '@/types/entities';
-import { inject, type Ref } from 'vue';
+import { PaymentTrackingMode, type Group } from '@/types/entities';
+import { computed, inject, type Ref } from 'vue';
 
 export function useGroupContext() {
 	const group = inject<Ref<Group>>(CONTEXT_KEY.GROUP, { value: {} } as Ref<Group>);
+	const isAccountantMode = computed(
+		() => group.value.paymentTrackingMode === PaymentTrackingMode.Accountant,
+	);
 
-	return { group };
+	return { group, isAccountantMode };
 }
