+++ src/features/group-detail/members/AccountingMaker.vue
@@ -1,13 +1,16 @@
 <script setup lang="ts">
 import Flex from '@/components/ui/Flex.vue';
+import { useGroupContext } from '../hooks/useGroupContext';
 import AccountingIcon from './AccountingIcon.vue';
 
 defineProps<{ show?: boolean }>();
 defineOptions({ inheritAttrs: false });
+
+const { isAccountantMode } = useGroupContext();
 </script>
 
 <template>
-	<div class="relative shrink-0" v-bind="$attrs" v-if="show">
+	<div class="relative shrink-0" v-bind="$attrs" v-if="show && isAccountantMode">
 		<slot></slot>
 		<Flex
 			data-tip="Kế toán"
